Refactor web workspace session hook
